LPS-55533 Extract the enum
