+++ modules/apps/wiki/wiki-item-selector-web/src/META-INF/resources/init.jsp
@@ -19,10 +19,10 @@
 <%@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 <%@ taglib uri="http://liferay.com/tld/ui" prefix="liferay-ui" %>
 
-<%@ page import="com.liferay.portal.kernel.dao.search.SearchContainer" %><%@
+<%@ page import="com.liferay.item.selector.taglib.ReturnType" %><%@
+page import="com.liferay.portal.kernel.dao.search.SearchContainer" %><%@
 page import="com.liferay.portal.kernel.util.ParamUtil" %><%@
 page import="com.liferay.portlet.PortletURLUtil" %><%@
-page import="com.liferay.taglib.ui.ItemSelectorBrowserTag" %><%@
 page import="com.liferay.wiki.item.selector.web.WikiAttachmentItemSelectorCriterion" %><%@
 page import="com.liferay.wiki.item.selector.web.WikiAttachmentItemSelectorView" %><%@
 page import="com.liferay.wiki.model.WikiPage" %><%@
